# Conflicts:
#	Assets/Scenes/Main_Scenes.unity

+++ Assets/Minh/Buff_Effect/BuffEffectForPerk/BuffStrengthByItems.cs
@@ -42,7 +42,7 @@ public class BuffStrengthByItems : IBuffEffect
     {
         if (GamePlayController.Instance.ItemController.IsPickupItemSuccess == false)
         {
-            player.Stats.ChangeStrength(Value);
+            player.Stats.ChangeStrength((int)Value);
         }
     }
