+++ Assets/Minh/Item_Script/SpecificBuffEffect/WoodenBracelet.cs
@@ -9,7 +9,7 @@ public class WoodenBracelet : BuffItem
 
     public override void Buff(Player player)
     {
-        player.Stats.ChangeStrength(buffVal * 0.1f);
+        player.Stats.ChangeStrength((int)(buffVal * 0.1f));
         player.Stats.ChangeShield(buffVal);
     }
